+++ tests/SchemaTest.php
@@ -5,44 +5,31 @@ declare(strict_types=1);
 namespace Yiisoft\Db\Mssql\Tests;
 
 use PHPUnit\Framework\Attributes\DataProviderExternal;
-use Throwable;
 use Yiisoft\Db\Command\CommandInterface;
 use Yiisoft\Db\Connection\ConnectionInterface;
 use Yiisoft\Db\Driver\Pdo\PdoConnectionInterface;
-use Yiisoft\Db\Exception\Exception;
-use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
 use Yiisoft\Db\Mssql\Schema;
 use Yiisoft\Db\Mssql\Tests\Provider\SchemaProvider;
 use Yiisoft\Db\Mssql\Tests\Support\TestTrait;
 use Yiisoft\Db\Schema\Column\ColumnInterface;
-use Yiisoft\Db\Schema\SchemaInterface;
 use Yiisoft\Db\Schema\TableSchemaInterface;
 use Yiisoft\Db\Tests\Common\CommonSchemaTest;
 use Yiisoft\Db\Tests\Support\DbHelper;
 
 /**
  * @group mssql
- *
- * @psalm-suppress PropertyNotSetInConstructor
  */
 final class SchemaTest extends CommonSchemaTest
 {
     use TestTrait;
 
-    /**
-     * @dataProvider \Yiisoft\Db\Mssql\Tests\Provider\SchemaProvider::columns
-     */
+    #[DataProviderExternal(SchemaProvider::class, 'columns')]
     public function testColumns(array $columns, string $tableName): void
     {
         parent::testColumns($columns, $tableName);
     }
 
-    /**
-     * @throws Exception
-     * @throws InvalidConfigException
-     * @throws Throwable
-     */
     public function testFindUniquesIndexes(): void
     {
         $db = $this->getConnection(true);
@@ -62,9 +49,6 @@ final class SchemaTest extends CommonSchemaTest
         $this->assertSame('dbo', $schema->getDefaultSchema());
     }
 
-    /**
-     * @throws NotSupportedException
-     */
     public function testGetSchemaNames(): void
     {
         $db = $this->getConnection(true);
@@ -86,39 +70,25 @@ final class SchemaTest extends CommonSchemaTest
         $this->assertSame(['animal_view'], $schema->getViewNames('dbo', true));
     }
 
-    /**
-     * @dataProvider \Yiisoft\Db\Mssql\Tests\Provider\SchemaProvider::constraints
-     *
-     * @throws Exception
-     */
+    #[DataProviderExternal(SchemaProvider::class, 'constraints')]
     public function testTableSchemaConstraints(string $tableName, string $type, mixed $expected): void
     {
         parent::testTableSchemaConstraints($tableName, $type, $expected);
     }
 
-    /**
-     * @dataProvider \Yiisoft\Db\Mssql\Tests\Provider\SchemaProvider::constraints
-     *
-     * @throws Exception
-     */
+    #[DataProviderExternal(SchemaProvider::class, 'constraints')]
     public function testTableSchemaConstraintsWithPdoLowercase(string $tableName, string $type, mixed $expected): void
     {
         parent::testTableSchemaConstraintsWithPdoLowercase($tableName, $type, $expected);
     }
 
-    /**
-     * @dataProvider \Yiisoft\Db\Mssql\Tests\Provider\SchemaProvider::constraints
-     *
-     * @throws Exception
-     */
+    #[DataProviderExternal(SchemaProvider::class, 'constraints')]
     public function testTableSchemaConstraintsWithPdoUppercase(string $tableName, string $type, mixed $expected): void
     {
         parent::testTableSchemaConstraintsWithPdoUppercase($tableName, $type, $expected);
     }
 
-    /**
-     * @dataProvider \Yiisoft\Db\Mssql\Tests\Provider\SchemaProvider::tableSchemaWithDbSchemes
-     */
+    #[DataProviderExternal(SchemaProvider::class, 'tableSchemaWithDbSchemes')]
     public function testTableSchemaWithDbSchemes(string $tableName, string $expectedTableName): void
     {
         $db = $this->getConnection();
@@ -141,24 +111,9 @@ final class SchemaTest extends CommonSchemaTest
             )
             ->willReturn($commandMock);
         $schema = new Schema($mockDb, DbHelper::getSchemaCache());
-        $schema->getTablePrimaryKey($tableName);
-    }
+        $schema->getTablePrimaryKey($tableName, true);
 
-    public function withIndexDataProvider(): array
-    {
-        return [
-            ...parent::withIndexDataProvider(),
-            [
-                'indexType' => SchemaInterface::INDEX_CLUSTERED,
-                'indexMethod' => null,
-                'columnType' => 'varchar(16)',
-            ],
-            [
-                'indexType' => SchemaInterface::INDEX_NONCLUSTERED,
-                'indexMethod' => null,
-                'columnType' => 'varchar(16)',
-            ],
-        ];
+        $db->close();
     }
 
     public function testNotConnectionPDO(): void
@@ -190,8 +145,8 @@ final class SchemaTest extends CommonSchemaTest
     }
 
     #[DataProviderExternal(SchemaProvider::class, 'resultColumns')]
-    public function testGetResultColumn(ColumnInterface|null $expected, array $info): void
+    public function testGetResultColumn(ColumnInterface|null $expected, array $metadata): void
     {
-        parent::testGetResultColumn($expected, $info);
+        parent::testGetResultColumn($expected, $metadata);
     }
 }