+++ CHANGELOG.md
@@ -51,6 +51,7 @@
 - Enh #375: Adapt to `Param` refactoring in `yiisoft/db` package (@vjik)
 - Enh #376, #379: Adapt to conditions refactoring in `yiisoft/db` package (@vjik)
 - Enh #380: Remove support dots in table names (@Tigrov)
+- Enh #383: Refactor `TableSchema` and `Schema` classes (@Tigrov)
 
 ## 1.2.0 March 21, 2024